Refactor: [#17] 서버 도메인 기준에 맞춰 API 리팩토링
- loginAPI, onboardingAPI 삭제
- 명세서에 구분된 도메인에 맞춰 api 생성 후 이동.
