+++ here-draw-ios/Feature/Onboarding/OnboardingViewModel.swift
@@ -10,7 +10,7 @@ import UIKit
 class OnboardingViewModel {
     // TODO: 닉네임 중복 검사
     func doubleCheckNickname(nickname: String, onCompletion: @escaping (Bool, String) -> Void) {
-        OnboardingAPI.checkNickname(nickname: nickname) { response in
+        UserAPI.checkNickname(nickname: nickname) { response in
             if let result = response.result {
                 onCompletion(true, result)
             }
@@ -22,7 +22,7 @@ class OnboardingViewModel {
     
     // TODO: 닉네임 설정
     func setNickname(nickname: String, onCompletion: @escaping (Bool, String) -> Void) {
-        OnboardingAPI.registerNickname(parameters: NicknameRequest(nickname: nickname)) { response in
+        UserAPI.registerNickname(parameters: NicknameRequest(nickname: nickname)) { response in
             if let result = response.result {
                 UserDefaults.standard.set(nickname, forKey: "nickname")
                 onCompletion(true, result)
